Use modern style for script parameters, remove logging options
The logging related options are useful for development, but not relevant
for the end user. Hence they are now hard-coded in the macro instead of
presented in the dialog.

If required, anyone can open the macro code from the package by hitting the
"Source" button in the search results and adjust the corresponding lines.