Ensure to convert to 8-bit in a reasonable fashion
Unlike what we had assumed, the "resetMinAndMax()" command doesn't set the
display range to min and max of the current bit depth but rather to the min and
max from the actual values. This was causing a high offset (many pixels with
value zero) after converting to 8 bit, so we rather calculate the display range
now from the actual values before converting to 8-bit.